requirements.txt
GitPython==3.1.3
PyYAML
pydantic
